requirements.txt
Flask
mysql-connector-python
gunicorn
python-dotenv
